Refactor: split commits related functions to submodules
